+++ src/context.rs
@@ -1,0 +1,7 @@
+use crate::{config::Config, db::core::PgConnection};
+
+#[derive(Clone)]
+pub struct Context {
+    pub config: Config,
+    pub db: PgConnection,
+}
