--- src/utils/mod.rs
@@ -1,2 +1,0 @@
-pub mod core;
-pub mod jwt;
